+++ lib/element.ts
@@ -423,9 +423,7 @@ export abstract class AstElement extends AstNode {
 	 * @throws `SyntaxError` 未定义的伪选择器
 	 */
 	#matches(step: SelectorArray): boolean {
-		const {
-				parentNode, type, name, childNodes, link, fixed, constructor: {name: tokenName},
-			} = this as AstElement & {link?: string | Title},
+		const {parentNode, type, name, childNodes, link, fixed} = this as AstElement & {link?: string | Title},
 			children = parentNode?.children,
 			childrenOfType = children?.filter(({type: t}) => t === type),
 			siblingsCount = children?.length ?? 1,
@@ -476,7 +474,7 @@ export abstract class AstElement extends AstNode {
 					case ':read-write':
 						return !fixed;
 					case ':invalid':
-						return type === 'table-inter' || tokenName === 'HiddenToken';
+						return type === 'table-inter' || type === 'image-parameter' && name === 'invalid';
 					case ':required':
 						return this.#isProtected() === true;
 					case ':optional':