+++ lib/text.ts
@@ -130,7 +130,6 @@ export class AstText extends AstNode {
 						startCol = lines.length === 1 ? left + line.length : line.length,
 						{0: char, length} = error,
 						endIndex = startIndex + length,
-						end = char === '}' || char === ']' ? endIndex + 1 : startIndex + 49,
 						rootStr = String(root),
 						nextChar = rootStr[endIndex],
 						previousChar = rootStr[startIndex - 1],
@@ -156,7 +155,6 @@ export class AstText extends AstNode {
 						endLine: startLine,
 						startCol,
 						endCol: startCol + length,
-						excerpt: rootStr.slice(Math.max(0, end - 50), end),
 					};
 				}).filter(Boolean) as LintError[];
 		}