+++ src/magicLink.ts
@@ -114,7 +114,6 @@ export class MagicLinkToken extends syntax(Token) {
 					endLine: startLine,
 					startCol,
 					endCol: startCol + s.length,
-					excerpt: data.slice(Math.max(0, index! - 25), index! + 25),
 				};
 			}));
 		}