+++ src/nowiki/quote.ts
@@ -16,10 +16,8 @@ export class QuoteToken extends syntax(NowikiBaseToken, /^(?:'{5}|'''?)$/u) {
 			message = Parser.msg('lonely "$1"', `'`),
 			errors: LintError[] = [];
 		let refError: LintError | undefined;
-		let wikitext: string | undefined;
 		if (previousSibling?.type === 'text' && previousSibling.data.endsWith(`'`)) {
 			refError = generateForSelf(this, {start}, message);
-			wikitext = String(this.getRootNode());
 			const {startIndex: endIndex, startLine: endLine, startCol: endCol} = refError,
 				[{length}] = /(?<!')'+$/u.exec(previousSibling.data)!,
 				startIndex = start - length;
@@ -30,12 +28,10 @@ export class QuoteToken extends syntax(NowikiBaseToken, /^(?:'{5}|'''?)$/u) {
 				startCol: endCol - length,
 				endLine,
 				endCol,
-				excerpt: wikitext.slice(startIndex, startIndex + 50),
 			});
 		}
 		if (nextSibling?.type === 'text' && nextSibling.data.startsWith(`'`)) {
 			refError ??= generateForSelf(this, {start}, message);
-			wikitext ??= String(this.getRootNode());
 			const {endIndex: startIndex, endLine: startLine, endCol: startCol} = refError,
 				[{length}] = /^'+/u.exec(nextSibling.data)!,
 				endIndex = startIndex + length;
@@ -46,7 +42,6 @@ export class QuoteToken extends syntax(NowikiBaseToken, /^(?:'{5}|'''?)$/u) {
 				startLine,
 				startCol,
 				endCol: startCol + length,
-				excerpt: wikitext.slice(Math.max(0, endIndex - 50), endIndex),
 			});
 		}
 		return errors;